Fix main load failure: rebuild lib/ammitto.rb on the #16 source layout
The #12 rebase kept the pre-#16 requires (sources/*_source.rb), which #16
renamed to sources/<code>/source.rb — require 'ammitto' raised LoadError and
the suite ran 0 examples while every data-repo cron installs main unlocked.
Rebuilds lib/ammitto.rb on the #16 layout with #12's constant qualification
re-applied, ports the five #12 sources into the source.rb convention, and
points sources/cn.rb at cn/source. cli.rb requires the core library again
(#16's XML-adapter bootstrap; also keeps the standalone version command
working).
