+++ lib/ammitto.rb
@@ -31,21 +31,21 @@ require_relative 'ammitto/models'
 # Sources
 require_relative 'ammitto/sources/registry'
 require_relative 'ammitto/sources/base_source'
-require_relative 'ammitto/sources/eu_source'
-require_relative 'ammitto/sources/un_source'
-require_relative 'ammitto/sources/us_source'
-require_relative 'ammitto/sources/wb_source'
-require_relative 'ammitto/sources/uk_source'
-require_relative 'ammitto/sources/au_source'
-require_relative 'ammitto/sources/ca_source'
-require_relative 'ammitto/sources/ch_source'
-require_relative 'ammitto/sources/cn_source'
-require_relative 'ammitto/sources/ru_source'
-require_relative 'ammitto/sources/tr_source'
-require_relative 'ammitto/sources/nz_source'
-require_relative 'ammitto/sources/jp_source'
-require_relative 'ammitto/sources/eu_vessels_source'
-require_relative 'ammitto/sources/un_vessels_source'
+require_relative 'ammitto/sources/eu/source'
+require_relative 'ammitto/sources/un/source'
+require_relative 'ammitto/sources/us/source'
+require_relative 'ammitto/sources/wb/source'
+require_relative 'ammitto/sources/uk/source'
+require_relative 'ammitto/sources/au/source'
+require_relative 'ammitto/sources/ca/source'
+require_relative 'ammitto/sources/ch/source'
+require_relative 'ammitto/sources/cn/source'
+require_relative 'ammitto/sources/ru/source'
+require_relative 'ammitto/sources/tr/source'
+require_relative 'ammitto/sources/nz/source'
+require_relative 'ammitto/sources/jp/source'
+require_relative 'ammitto/sources/eu_vessels/source'
+require_relative 'ammitto/sources/un_vessels/source'
 
 # Client
 require_relative 'ammitto/client/api_client'
