--- lib/ammitto/sources/nz_source.rb
@@ -1,36 +1,0 @@
-# frozen_string_literal: true
-
-# NZ source models (Lutaml::Model)
-require_relative 'nz/sanctions_list'
-require_relative 'nz/transformer'
-
-# Ammitto top-level namespace
-module Ammitto
-  # NzSource handles New Zealand (MFAT) sanctions data
-  #
-  # @example
-  #   source = NzSource.new
-  #   data = source.load_data
-  #   results = source.search("name", data)
-  #
-  class NzSource < BaseSource
-    # @return [Symbol] the source code
-    def code
-      :nz
-    end
-
-    # @return [Authority] the issuing authority
-    def authority
-      @authority ||= Authority.find('nz')
-    end
-
-    # Get the original source endpoint
-    # @return [String] the official source URL
-    def original_api_endpoint
-      'https://www.mfat.govt.nz/en/countries-and-regions/europe/ukraine/russian-invasion-of-ukraine/sanctions/'
-    end
-  end
-
-  # Register the source
-  Registry.register(:nz, NzSource)
-end
